requirements.txt
gunicorn
flask
tensorflow==2.12.0
pillow
numpy
protobuf==3.20.*